Added the ability to filter players by tribe to the Data Service.
This is going to be how multiple teams can exist on the same server. I ran an experiment with trying to keep players and history in the same document... but it didn't work so well.

The idea is a player can only have one tribe. There might be a future where tribes subdivide into clans... this would be for easily swapping players between related teams in the UI.
- RMUR