Started refactoring the index page and CouplingController.
My current rule of thumb is: methods called from html should be named according to where they are being called from. This is so that the view doesn't have to know very much about what the event will do, just that it exists.  This also makes it clearer on the controller that the primary purpose of the method is to bind to the event.
- RMUR